backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy[asyncio]==2.0.40
sqlmodel==0.0.24
asyncpg==0.30.0
databases==0.9.0
alembic==1.13.3
pymongo==4.12.0
motor==3.7.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
slowapi==0.1.9
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1